refactor(lib): inline namespaced functions
the other approach leads to pitfalls with JSII
